(chore) move isValidNewList to helper file
